backend/requirements.txt
fastapi
uvicorn
diffusers[torch]
transformers
torch
pillow 